rand_xorshift reorganisation, courtesy of @newpavlov
This copies changes from the separate small-rngs repo
