+++ rand_xorshift/Cargo.toml
@@ -5,8 +5,8 @@ authors = ["The Rand Project Developers", "The Rust Project Developers"]
 license = "MIT/Apache-2.0"
 readme = "README.md"
 repository = "https://github.com/rust-random/rand"
-documentation = "https://docs.rs/rand_isaac"
-homepage = "https://crates.io/crates/rand_isaac"
+documentation = "https://docs.rs/rand_xorshift"
+homepage = "https://crates.io/crates/rand_xorshift"
 description = """
 Xorshift random number generator
 """
